# services/api/gen/openapi/openapi.bundle.yaml
# Code generated by scripts/openapi-bundle from api/openapi/; DO NOT EDIT.
# Edit the source files under api/openapi/ and run `make gen-openapi`.
components:
  parameters:
    UserId:
      in: path
      name: id
      required: true
      schema:
        format: int64
        minimum: 1
        type: integer
  responses:
    BadRequest:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: The request is malformed.
    Conflict:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: The request conflicts with current state.
    InternalError:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: Unexpected server error.
    NotFound:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: The resource does not exist.
    Unauthorized:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: Missing or invalid credentials.
    Unavailable:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: A dependency is down.
    ValidationFailed:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: The request failed validation; `fields` lists the problems.
  schemas:
    ChangePasswordRequest:
      properties:
        currentPassword:
          format: password
          type: string
        newPassword:
          format: password
          maxLength: 256
          minLength: 12
          type: string
      required:
        - currentPassword
        - newPassword
      type: object
    CreateUserRequest:
      properties:
        email:
          example: ada@example.com
          format: email
          type: string
        name:
          example: Ada Lovelace
          maxLength: 120
          type: string
        password:
          format: password
          maxLength: 256
          minLength: 12
          type: string
      required:
        - email
        - name
        - password
      type: object
    Error:
      properties:
        code:
          description: HTTP status.
          example: 404
          type: integer
        error:
          description: Machine-readable error code.
          example: RESOURCE_NOT_FOUND
          type: string
        fields:
          additionalProperties:
            type: string
          description: Field name to validation message.
          example:
            password: must be at least 12 characters
          type: object
        message:
          example: user not found
          type: string
        request_id:
          example: 8f1c2a9d-3b7e-4f5a-9c1d-2e3f4a5b6c7d
          type: string
      required:
        - error
        - message
        - code
      type: object
    Health:
      properties:
        status:
          enum:
            - ok
          type: string
      required:
        - status
      type: object
    Readiness:
      properties:
        checks:
          additionalProperties:
            type: string
          example:
            db: ok
          type: object
        status:
          enum:
            - ok
          type: string
      required:
        - status
        - checks
      type: object
    UpdateUserRequest:
      properties:
        name:
          example: Ada King
          maxLength: 120
          type: string
      required:
        - name
      type: object
    User:
      properties:
        createdAt:
          format: date-time
          type: string
        email:
          example: ada@example.com
          format: email
          type: string
        emailVerified:
          type: boolean
        id:
          example: 42
          format: int64
          type: integer
        name:
          example: Ada Lovelace
          maxLength: 120
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - id
        - email
        - name
        - emailVerified
        - createdAt
        - updatedAt
      type: object
    UserList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/User'
          type: array
        total:
          description: Number of users matching the filter.
          example: 42
          type: integer
      required:
        - items
        - total
      type: object
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
    cookieAuth:
      in: cookie
      name: access_token
      type: apiKey
info:
  contact:
    name: scaffold maintainers
    url: https://github.com/bete7512/scaffold
  description: Errors are JSON `Error` objects; every response carries an `X-Request-Id` header.
  license:
    name: MIT
  title: scaffold API
  version: 0.1.0
openapi: 3.0.3
paths:
  /healthz:
    get:
      description: Reports that the process is up and serving; never checks dependencies, so a database blip does not get the task killed.
      operationId: getHealthz
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
          description: Process is alive.
      security: []
      summary: Liveness probe
      tags:
        - system
  /readyz:
    get:
      description: Pings every hard dependency (currently Postgres) and returns 503 with per-check detail when any is down, so the load balancer takes this replica out of rotation.
      operationId: getReadyz
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Readiness'
          description: All dependency checks passed.
        "503":
          $ref: '#/components/responses/Unavailable'
      security: []
      summary: Readiness probe
      tags:
        - system
  /v1/me:
    get:
      description: Returns the profile of the user identified by the bearer token or access_token cookie; 401 when neither is present or valid.
      operationId: getMe
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: The authenticated user.
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalError'
      security:
        - bearerAuth: []
        - cookieAuth: []
      summary: Get the current user
      tags:
        - users
  /v1/users:
    get:
      description: Paginated with limit and offset, sorted by `sort_by` (default id) and `sort_dir` (default asc).
      operationId: listUsers
      parameters:
        - in: query
          name: limit
          schema:
            default: 20
            maximum: 100
            minimum: 1
            type: integer
        - in: query
          name: offset
          schema:
            default: 0
            minimum: 0
            type: integer
        - description: Case-insensitive match on email or name.
          in: query
          name: search
          schema:
            type: string
        - in: query
          name: sort_by
          schema:
            default: id
            enum:
              - id
              - email
              - name
              - createdAt
            type: string
        - in: query
          name: sort_dir
          schema:
            default: asc
            enum:
              - asc
              - desc
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
          description: A page of users.
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "422":
          $ref: '#/components/responses/ValidationFailed'
        "500":
          $ref: '#/components/responses/InternalError'
      security:
        - bearerAuth: []
        - cookieAuth: []
      summary: List users
      tags:
        - users
    post:
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: User created.
        "400":
          $ref: '#/components/responses/BadRequest'
        "409":
          $ref: '#/components/responses/Conflict'
        "422":
          $ref: '#/components/responses/ValidationFailed'
        "500":
          $ref: '#/components/responses/InternalError'
      security: []
      summary: Register a user
      tags:
        - users
  /v1/users/{id}:
    delete:
      operationId: deleteUser
      responses:
        "204":
          description: Deleted.
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalError'
      security:
        - bearerAuth: []
        - cookieAuth: []
      summary: Delete a user
      tags:
        - users
    get:
      operationId: getUser
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: The user.
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalError'
      security:
        - bearerAuth: []
        - cookieAuth: []
      summary: Get a user
      tags:
        - users
    parameters:
      - $ref: '#/components/parameters/UserId'
    patch:
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: The updated user.
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: '#/components/responses/ValidationFailed'
        "500":
          $ref: '#/components/responses/InternalError'
      security:
        - bearerAuth: []
        - cookieAuth: []
      summary: Update a user's profile
      tags:
        - users
  /v1/users/{id}/password:
    put:
      description: Requires the current password. Signs the user out of every session.
      operationId: changeUserPassword
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
        required: true
      responses:
        "204":
          description: Password changed.
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: '#/components/responses/ValidationFailed'
        "500":
          $ref: '#/components/responses/InternalError'
      security:
        - bearerAuth: []
        - cookieAuth: []
      summary: Change a user's password
      tags:
        - users
servers:
  - description: Local development (`make dev`)
    url: http://localhost:8080
tags:
  - description: Liveness / readiness probes used by the load balancer and orchestrator.
    name: system
  - description: User registration and profile.
    name: users